+++ Sources/ZenPackageMetadata/ZenPackageMetadata.swift
@@ -6,5 +6,5 @@
 //
 
 public enum ZenPackageMetadata {
-    public static let version = "2.0.1"
+    public static let version = "2.1.0"
 }